requirements.txt
# Core trading dependencies
kiteconnect>=4.0.0

# Environment configuration
python-dotenv>=1.0.0

# FastAPI server dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0

# HTTP requests (for testing and client examples)
requests>=2.28.0

# Additional utilities
python-multipart>=0.0.6
